nccl_ep: fp8 dispatch (EXTERN) post-review fixes
Three changes per review feedback:

1. inScalesBuf type: const float* -> const uint8_t* throughout the LL
   dispatch path (host wrapper, kernel param, explicit instantiations).
   Callers pass raw scale bytes; LL dispatch byte-copies verbatim.
   castAndWriteToSendBuf absorbs the EXTERN copy path (was a separate
   copyExternFP8ToSendBuf); the EXTERN variant selects int2* src and
   byte-copies both payload and scales in-place.

2. kExternFP8 -> kExternQuant: dispatch is format-agnostic; the name
   should not imply FP8. srcDataFP8Extern -> srcDataExternQuant.

3. HT sf_bytes_per_token: replace SCALE_ELEM_BYTES (template) +
   scale_block_size (runtime) with a single pre-computed value passed
   from host. Propagated through dispatch_config_t, N2N/G2S/S2G warp
   groups, dispatch_kernel_impl, hybridep_adapter, JIT source/launch
   (variant suffix _sb<N>_se<N> -> _sf<N>), and nccl_ep.cc call site.

All four test configurations pass: HT FP8 EXTERN, LL FP8 EXTERN, HT
BF16, LL BF16.

+++ nccl_ep/include/common.hpp
@@ -167,7 +167,7 @@ void dispatch(const void* inData,
               const float* inTopkWeights,    // rank-major: written into dispatch message header
               void* outDataBuf,
               void* outScalesBuf,
-              const float* inScalesBuf, // non-null for LL FP8 EXTERN (inputs->scales)
+              const uint8_t* inScalesBuf, // non-null for LL EXTERN quant (inputs->scales)
               int* outSrcInfo,
               int* outRecvRankCounter,          // rank-major: RECV_RANK_COUNTER_DEVICE [nRanks]; nullptr for expert-major
               int64_t* outLayout,
